requirements.txt
# AI Agent Requirements - Minimal and focused
# For your Email_Codes AI agent project

# Core OpenAI API
openai>=1.3.0

# DSPy for signatures, modules, optimizers
dspy-ai>=2.5.0

# Web scraping (for tools.py)
selenium>=4.15.0
beautifulsoup4>=4.12.0

# FastAPI for API endpoints
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6

# Configuration and environment
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# HTTP requests
requests>=2.31.0
httpx>=0.25.0

# LangChain orchestration (pin compatible versions)
langchain==0.3.8
langchain-community==0.3.8
langchain-openai==0.2.7

# Wikipedia search helper
wikipedia>=1.4.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
